backend/pyproject.toml
[project]
name = "comic-engine-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "torch>=2.2.0",
    "transformers>=4.40.0",
    "pillow>=10.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "httpx>=0.27.0", "anyio[trio]>=4.0.0"]

[tool.setuptools.packages.find]
include = ["comic_engine*"]
